requirements-production.txt
# MapMyStandards Production Requirements (Optimized for Railway)
# Python 3.9+ Required

# Core Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.0
gunicorn==21.2.0

# API Dependencies  
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.4.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
email-validator==2.1.0
cryptography==41.0.7

# Payment Processing
stripe==12.4.0

# Email Services
postmarker==1.0  # Postmark Python library (primary)
sendgrid==6.11.0  # SendGrid as fallback
aiosmtplib==3.0.1
emails==0.6  # For email templates

# HTTP Client
httpx==0.28.1  # Updated to match pyautogen requirement
aiohttp==3.9.1

# AWS S3 Storage
boto3==1.34.0
botocore==1.34.0

# AI/LLM Services (Required for startup)
# Bumped OpenAI to a newer SDK compatible with httpx 0.28.x to avoid 'proxies' init errors
openai==1.40.6
anthropic==0.39.0

# Data Processing (Minimal)
pandas==2.0.3
numpy==1.24.3
pyyaml==6.0.2

# Document Processing (Essential only)
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pypdf==3.17.0
python-magic==0.4.27
chardet==5.2.0
Pillow==10.1.0  # For image processing
reportlab==4.2.0  # For PDF generation in report_generation

# Pinecone (without heavy ML deps)
pinecone-client==3.0.0

# Lightweight ML utilities used by analytics endpoints
scikit-learn==1.3.2

# Production essentials
redis==4.6.0
celery[redis]==5.3.4

# CORS Support
flask-cors==6.0.1

# Additional utilities
python-jose[cryptography]==3.3.0  # For JWT handling
itsdangerous==2.1.2  # For secure sessions
markupsafe==2.1.3  # For safe HTML
bleach==6.1.0  # For HTML sanitization

# Agent orchestration (optional but included for full functionality)
pyautogen==0.9.0  # For multi-agent workflows
prometheus-client==0.20.0  # Metrics exposition